pyproject.toml
[tool.poetry]
name = "arcade_notion"
version = "0.1.1"
description = "Notion integration to search, create, update, and delete databases/pages."
authors = ["Sergio Serrano <sdserranog@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
arcade-ai = "^0.1.1"
requests = "^2.31.0"
loguru = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"